Write down the path for aliases.
The path that Crubit infers only works by a seriously deranged coincidence (it Just So Happens that it both redundantly adds the namespace name, and omits the crate name, evening it out), and this blocks unknown commit.

After this CL, we'll use the correct crate path, and it will stay _constant_ even if we then change it so that Crubit is aware of the correct crate that this lives in.

PiperOrigin-RevId: 783540866
Change-Id: I54661365feccefd1124fa82d3e88d703b28ec7d9