Nix the phrasing "trivially relocatable".
We originally assumed, back in 2021, that trivially relocatable types would have a relocation operation which is "trivial" in the same sense that a trivial copy is "trivial". The current wording in C++26 doesn't go this way, and as a result is not useful to discuss in interop documentation. At least as of right now, in C++26, a trivially relocatable type may or may not be Rust-movable, and we need a separate type trait for trivial relocatability.

We should "XY problem" this. We don't actually care about the C++ term "trivially relocatable". We care about the term we can define, "Rust-movable". This avoids weird refactorings as C++ terminology changes, and brings our documentation up to date.

PiperOrigin-RevId: 783834033
Change-Id: I64d3736bd0b8ee335ba54b60deed4213e42736bd