Use only supported features for rs_bindings_from_cc golden tests
The previous use of experimental features in all goldens made the internal Crubit
development experience different from that of end users. This change makes it
easier to track what features are being released to users. It is still possible
to re-enable experimental on a case-by-case basis for tests that specifically
exercise experimental features.

PiperOrigin-RevId: 785484632
Change-Id: I4c2d122e5bac41141c096fdae6be47746e8abc56