Use `super_let` so that `ctor.rs` looks less dialectical.[^1]
Trying to construct a ever so slightly tweaked version of ctor.rs, and its public docs, that are, you know, approachable. Even before this CL, `Box::emplace` was an exact mirror of `Box::pin`. After this CL, `emplace!()` is an exact mirror of `std::pin!()`. And so, rather than memorizing a universe of things, we can, with a few more tweaks like this, try to characterize Crubit as just a generalized version of what Rust already has with its value semantics.

[^1]: (Or are we making it _more_ dialectical? Maybe it's with these small quantitative changes that we can produce a qualitative change. 🤔)

PiperOrigin-RevId: 785537216
Change-Id: Ic598f85cf61b2ae79f623068e9ca483506e46ae6