Add a `crubit/support/annotations.swig`.
This allows users to work around the annotations (which don't parse in SWIG) by including the swig dependency, probably. We don't need to maintain this except when it comes up in the targets that actually have SWIG bindings.

Since this isn't bindings for anything, I didn't place it in a language subdirectory. It can be used by any SWIG user, AFAIK.

PiperOrigin-RevId: 788095037
Change-Id: I74ea069f89e97bb9536f4673436328c1e93f57ef