Don't bridge tuple fields except by value.
Before this CL, the test would fail to compile. :'(

NOTE: A very first attempt at this CL did not introduce `TypeLocation::NestedBridgeable`, but this (in retrospect, obviously) would break nested tuples -- a nested tuples is not in parameter or return position, but also is  still bridgeable, unlike e.g. struct fields, pointees.
PiperOrigin-RevId: 789414029
Change-Id: I34640b8244a44a52a02a0e735ae20cf63a40540f