Use `operator delete(void*, size_t)` in Rust, same as (our) C++.
Whether `delete x` uses `operator delete(void*, size_t)` or the unsized variant, if both exist, is not specified by the standard, but experimentally and in conversation, our delete uses the sized overload. We should probably make this configurable so that this works if someone switches it to work the other way, but I understand sized delete to be more efficient.

Surprisingly, ASan seems not to complain, even though it noted mismatch between aligned new and unaligned delete (https://github.com/google/crubit/commit/de7943e09fc01c181c04cead4c046804837b8568). My guess is we define this behavior? Or it's random chance it works out (e.g. sampling, or it happens to be that the size information is stored in the same place for both)? Need to actually do this properly and get a failing test so we detect if we get it wrong. :/

PiperOrigin-RevId: 791164411
Change-Id: I23dd874cf37c122298c15f3b58b8dc7c71e26f01