[NFC] Evidence collection: Move computation of overridden methods out of Emit.
This change is part of a broader effort to separate emitting evidence from a direct dependency on the AST. In this step, we add a map for overridden methods and then capture that map in the Emit function, rather than computing them on demand in `Emit`.

Note that a side benefit of the new, map-based approach is that it avoids O(N^2) behavior when walking the overridden method hierarchy. This benefit is probably quite small in practice, though. Also, it may be offset by the cost of eagerly building this map for the entire TU, when before we only accessed the data on demand.

Now:

- `getVirtualMethodOverrides` computes two indexes, one mapping to all overridden methods, and another mapping to overrides.

- `getAdditionalTargetsForVirtualMethod` uses a map for overridden methods, rather than computing the overriden methods on demand and it returns USRs rather than decls.

PiperOrigin-RevId: 791650444
Change-Id: I471a2a5dbf56017b09b9cf8f81d519aee6608081