If a file isn't `.inc`, and isn't owned by a target, don't generate bindings.
In particular, don't assign ownership to whoever includes it!

This allows us to block individual header files with names like `.h`, so that they don't get bindings, even if everything else in the target does, by just omitting them from the header list.

(The downside is, we might need a separate textual headers entry somewhere in the input.)

This isn't enough to fix b/408994283, because the error messages will be quite terrible without further work, and we would still need to actually stop passing target_args around. But it might be enough to allow individual header-level blocking in the vein of b/429567103 without shuffling the problem around to those that include the header. At least for the purpose of working around bugs.

PiperOrigin-RevId: 792227924
Change-Id: I0a1bc99af3c5817b2d2632222ddcde37c3602089