Stricter parameter type assertion.
Follow-up to b09e75b.

Since these are parameters, we cannot simply wrap them in an invariant type constructor. This change uses a trait-based approach to test for type equality.

Note that all existing test expectations are already correct. However, with this change, a test expecting a reference parameter type will now fail if the actual parameter is a pointer, which would pass today since function calls are coercion sites.

PiperOrigin-RevId: 792413385
Change-Id: I80c0a10ed2de15dd15002dc34d2ac600c9fd491a