Make `vector<T>` covariant on `T`.
I always hate that phrasing, it's up there with "monoid in the category of endofunctors". What I mean is: make it so that you can assign a `vector<&'static i32>` to a `vector<&i32>`.

There were two options here -- I could use an `Option<Nonnull<T>>`, but this is actually pretty painful. It's easier, albeit maybe more confusing, to use `*const T`.

PiperOrigin-RevId: 794169283
Change-Id: I0509200c9446989c4796b4286ba260541c5415c3