Use `non_prod` externally without diffs.
This is a clever Bazel pun: internally, we have `//buildenv/target:non_prod`, and `//third_party/crubit/stuff`. Externally, the `//third_party/crubit/` becomes just `//`, and so if we put a `buildenv/target` directory, it Just Works.

This avoids using copybara transforms, which would in this case be potentially non-reversible, breaking our ability to accept PRs on these files.

PiperOrigin-RevId: 796603671
Change-Id: Id4dd02896db21798443971ba37e59790e726f5ed