Catch more cases of inline functions.
Thunks for inline functions cannot be elided, as it's possible no TU will emit the symbol at all, leading to a link error.

When an inline function is declared but not defined, don't generate bindings at all. Such declarations are legal in C++ but cannot be ODR-used, and so we don't have any way to call them.

This CL adds definitions to a bunch of overly-simplified testcases that declared inline functions but did not define them, and are now caught be the second case above.

PiperOrigin-RevId: 797344811
Change-Id: I5dcc21135e3368803468069073e0b19d13156b9e