Specialize basic_string_view to string_view
This is a MASSIVE HACK but it fixes the bug. In order to explain why it's a hack, I first need to explain what the problem is.

# Problems

There are two fundamental problems with string_view and bridging:
1. The name `std::basic_string_view<...>` cannot be spelled with a `FullyQualifiedPath`. This is user facing _today_ and is what the bug refers to.
2. Each crate has its own `crate::__CcTemplateInst...` for the string_view inst, and these are incompatible with each other. This becomes problematic once you fix the first bug.

This CL intends to address _both_ of these issues.

# Exploration and solutions

To address the first problem, I inserted an "if the name == `std::basic_string_view<...`, use `std::string_view` instead". This let's us use `FullyQualifiedPath` to represent the name of the C++ type. One alternative I considered was that `FullyQualifiedPath` should be able to spell out templated types that are only written in cc files, but since this is such a special case, I didn't think it was worth it. At this point, the `CrubitAbiType` can spell itself in the generated cc file, but in Rust it will be spelled as `crate::__CcTemplateInst...`, leading us the the second problem.

string_view is a type alias, roughly:

```cpp
using string_view = basic_string_view<char, char_traits<char>>;
```

When Rust programs use `cc_std::std::string_view` directly, everything is fine because it aliases the `cc_std`'s template inst. But the fact is still that each crate still gets it's own `__CcTemplateInst...`, and when we bridge something like `std::optional<std::string_view>`, the alias is stripped away and it will bridge to `Option<crate::__CcTemplateInst...>`. Worst case, we would want this to be `Option<::cc_std::__CcTemplateInst...>`, best case it uses the alias (but that's impossible to recover without hardcoding).

The solution I implemented in this CL is dirty: I simply map the `owning_target` of the string_view `Record` to its `defining_target`. This is nasty and bad because now you have several `Item`s that kind of look the same but are not, and each target still gets it's own `__CcTemplateInstNSt3__u17basic_string_viewIDiNS_11char_traitsIDiEEEE`. However, it is good because whenever the `Item` is used (for example, in the type of something) it will appear as if it were the original template inst item from `cc_std`. So in the actual codegen, everything compiles and is good.

# Where do we go from here?

All we need to do is be able to unify _known_ template instantiations across build targets. `std::basic_string_view<char, std::char_traits<char>>` is a _known_ instantiation, and we know that cc_std is a target that creates that instantiation. To unify this inst, we should be able to recognize this instantiation if we're not in cc_std, and point to that one instead. This is like the Crubit version of what `extern template` does in C++11, but where we specify where the canonical inst lives.

Once we do this, it will also eliminate each target from getting their own string_view inst, which should fix other problems like b/436620437.

PiperOrigin-RevId: 797402551
Change-Id: I337c64eb55775ad201cb5a1c21f8159e98c52c49