Expose low-level evidence emitter directly to all call sites.
To sever dependency on the AST, callers must provide Evidence directly to the emitter, rather than passing AST references, SourceLocations, etc. as is done now. This CL makes `EvidenceEmitter` a `void(Evidence)` type and factors the wrapping functionality into a separate function, `wrappedEmit` that `Emit` callers now call explicitly. This sets the stage for incrementally migrating call sites to call the lower-level Emit function directly, avoiding `wrappedEmit` entirely.

However, the low-level `Emit` keeps the secondary role of propagating evidence along virtual-method chains. So, we reflect this functionality explicitly in the name to make clear to callers that it is not just emitting the evidence.

PiperOrigin-RevId: 797773458
Change-Id: Iee0b7a3cb6e89aade71713f91ed9428967d60d45