Document mangled_cc_name
I thought we could delete it but it turns out it's needed for generating unique upcast function names when an offset isn't available. I also tried to refactor `BaseClass` to have an either between offset OR mangled names, but that made generating the error trait for derived structs not getting a function impossible without really messy business.

PiperOrigin-RevId: 797919455
Change-Id: Ibd30830729f39158376d24ead15889be60b5332c