Apply a silly hack to improve `assert_ir_matches!` error messages sometimes
That macro uses `...` as a wildcard when matching token streams. Initially, it seems, `rustfmt` was willing to format the code snippets we use in tests that contain that token. And then at some point it started raising an error (quite reasonably), but we didn't notice because the checked-in tests pass and this only runs to produce an error message.

I think this is a bad solution. But I also think a good solution for this problem is not a small change. And this change helped me fix an internal issue involving one of these tests.

PiperOrigin-RevId: 798341739
Change-Id: Ife67152724b5152a178a6055c63f45c890124e69