Add C++ annotation `CRUBIT_UNSAFE_IGNORE_ATTR`
This change implements an initial subset of b/439606536. The new C++ annotation macro `CRUBIT_UNSAFE_IGNORE_ATTR(foo)` tells Crubit to ignore any attribute named `foo` on the same declaration for the purpose of identifying unknown attributes.

This annotation does not yet tell Crubit to ignore an attribute that it considers semantically meaningful. That's more involved because there are many places in the C++ side of `rs_bindings_from_cc` that directly query the AST for attributes.

PiperOrigin-RevId: 798361984
Change-Id: Ic71c4e6ea5655ad55c5bfa00d134ee2fee6ef0b3