Avoid untrackedError() in more cases of `diagnoseAssignmentLike` when RHS is missing nullability
In some cases we've filtered LHS and RHS to be pointer types ahead of time (diagnoseAssignment requires LHS isPointerExpr).

But, in some cases we have not. E.g., `diagnoseArgumentCompatibility` just goes through any of the args. In cases like foo(0), with the "RHS" being `0`, we wouldn't have computed an RHSNullability vector.

PiperOrigin-RevId: 799700782
Change-Id: I6e865cab4a448f472e08e0d44af5e4f70377b1cd