Create good error messages if the target is unknown, by using the filename.
This is dead code right now: we pretty much always know the target name. But once we stop passing it around in target_args, it's important to have good error messages.

I tested the change by actually _doing_ that, and looking at the goldens. So, for example, an error like:

> // //third_party/absl/types:span needs [//third_party/crubit/features:wrapper] for absl::Span<std::basic_string_view<char, std::char_traits<char>>> (::span::absl::RawSpanMut<cc_std::__CcTemplateInstNSt3__u17basic_string_viewIcNS_11char_traitsIcEEEE> is a template instantiation)

... will eventually become:

> // third_party/absl/types/span.h needs [//third_party/crubit/features:wrapper] for absl::Span<std::basic_string_view<char, std::char_traits<char>>> (::span::absl::RawSpanMut<cc_std::__CcTemplateInstNSt3__u17basic_string_viewIcNS_11char_traitsIcEEEE> is a template instantiation)

Meaning that, thanks to this CL, we can drop target assignment without a substantial drop in error message quality, in a followup change, after this version is released.

PiperOrigin-RevId: 799933332
Change-Id: If8f4ac357904d6427314b7eadd1d941225888153