Fix cc_deps attribute logic on additional_rust_srcs_for_crubit_bindings aspect hint
Before this change it failed because it required all targets passed in the cc_deps attribute to already have the Crubit aspect, and did not apply it to the target. Now it applies the crubit aspect to all targets in the label_list.

Also moved the provider to providers.bzl and moved a couple of utility functions to the only file that used them, otherwise we had circular dependencies in the BUILD graph.

PiperOrigin-RevId: 800059051
Change-Id: I512af8a825b3cff8a7b0e78264287505d0d8ec0a