Exclude the header-only `_cc_public_library_headers` target from getting Rust bindings.
I think this should work. It does mean that you can't get useful bindings for things that only depend on the target via this header-only target, but that seems inevitable, and easily fixed by changing what you depend on. In the meantime, it resolves the inherent ambiguity of who owns the declarations.

PiperOrigin-RevId: 800294864
Change-Id: Icbf10e40b2ea3c78ef74ecca43eb9e5ce7f4f42d