Separate evidence collection into 2 classes: summarization and collection.
This change explicitly separates evidence collection into 2 separate classes,
one for summarization of the AST and one for collection of evidence, based on
the summaries.  This new structure prepares the code for a subsequent
restructuring where summaries will be serialized (in protobufs), and then used
in multiple iterations of evidence collection.

PiperOrigin-RevId: 802535287
Change-Id: Ia9abac592da1e924554e4c0f042c8d63e208a04f