Add some checks before transferring pointer values to smart pointers
... to protect against possible crashes from type confusion.

If this happens though, the analysis does lose track and isn't well supported.
The most supported smart pointers look as close as possible to the std:: ones.

Previously `isPointerTypeConvertible` was relaxed and said
"If these aren't pointers to records, then just assume they're convertible.", I believe that's because it wasn't comparing the canonical types up front, and so fell through to that section. Now, we compare the canonical types earlier, and instead reject more cases later.

PiperOrigin-RevId: 803165018
Change-Id: I1357d832833f5d45786c1141b284ad6c96b15516