Make composable bridging work with existing Rust types.
This has two parts:

1. Delete is_same_abi check for ExistingRustType.

    AIUI it was written under the misapprehension that "is_same_abi" meant "is_same_layout", and required types to have the same layout. Since they always do, we can safely remove it.

2. Patch up the identifier parsing logic to allow `existing_rust_type.cc_name` to have namespaces in it. It's a really good thing I added a test, or I would've "fixed" the bug without actually making these types usable! :)

(Not a joke: I was about to mail without a test, thought, "nah, I should really test it". Very good thing I was not lazy.)

Note: there are two similar looking places you can get an error here, but the bug was not in the `.split(" :: ")` (though that could have failed if it happened somewhere else), but rather in the use of `existing_rust_type.cc_name` verbatim.

I did not look deeply into whether this is the right fix, but it does work. I also cleaned up the `.split()` Just In Case, since doing split by `"::"` and then a trim seems more robust than splitting by `" :: "`.

PiperOrigin-RevId: 805586051
Change-Id: I274802bcdb1795405934fd3f8c0df1822f633ec2