Fix crash while checking isPointerTypeConvertible in transferValue_SmartPointerAssignment
We were getting `underlyingRawPointerTypeFromSmartPointer(*SrcLoc)` but in some cases `SrcLoc` is null.

(a) Check if the getArg(1) is actually a glvalue before doing getStorageLocation(). In the example repro case, this actually asserts in an assert build -- before it returns nullptr. If it isn't a glvalue, instead use  getResultObjectLocation, which isn't nullptr.

(b) Defensively add a null check too -- though we aren't getting null anymore after (a).

PiperOrigin-RevId: 805755338
Change-Id: I4183704d5bc1acb625056026655ba7afe51a37f8