Add checks for TypeSourceInfo presence before dereferencing.
When collecting from declarations, we can't collect evidence from an annotated type if we can't find the type or an annotation.

When collecting from definitions, we can still collect evidence, but can only assume there is no existing annotation and we won't be able to actually annotate that type. We can still propagate an inference for that slot, though.

PiperOrigin-RevId: 806423166
Change-Id: Ib86c01839c8b28f45a3fba69493c090e056cba4f