Detect/reject overloaded operator delete and virtual destructors for std reimplementations.
The reimplementations of std::vector and std::unique_ptr bottom out at `::operator delete`. If a type overloads `operator delete`, then we'll have the completely wrong behavior, so this must be forbidden. If a type has a virtual destructor, then its operator delete is looked up dynamically, so this too must be forbidden. (Also, for separate reasons, virtual destructors with unique_ptr are difficult, as they require dynamically looking up the size, which there doesn't seem to be an API to do.)

For example, consider the following snippet:

```c++
struct X {
  static void operator delete(void*) {std::cerr << "Wow!\n";}
};

int main() {
    std::make_unique<X>();
}
```

You might reasonably expect this to print `Wow!`, because the destroyed unique_ptr will run `delete` for `T`, and `T` customizes `delete` via an overloaded `operator delete` to print `Wow!`.

However, the Rust reimplementation of `unique_ptr` (and destroying delete) does not have specialization or dispatch: it just runs the top-level deallocation function `::operator delete`, which sidesteps all of this. So it would be incorrect for `X`.

Therefore, we filter out these types.

With more advanced features in Rust, we could even warn if you ever happen to place one of these in a unique_ptr (e.g. we could try downcasting and check for success), but for now those don't kinds of casting APIs don't seem to exist.

PiperOrigin-RevId: 807403072
Change-Id: Id42de3da5e49dc6726bb920d3e00c0b4a91b7ae8