Allow for code to check for "is it std::allocator<T>" (and, for that matter, "is it std::default_delete<T>") without using features from `experimental`.
In particular, before this CL, you needed experimental features enabled in order to call `db.rs_type_kind(allocator_or_deleter)` to get the template specialization of e.g. `std::allocator<int>` and then query it. After this CL, we directly pull the `Record`. That's actually what it was doing before, too, it just went indirectly via rs_type_kind.

This CL is currently a no-op, but is necessary before we can add a separate (non-experimental) feature for unique_ptr and vector. Also, in crafting this CL, I broke things, and it can be hard to separate out expected vs unxpected diffs in the same CL, so I took the debugging step of extracting the supposed no-op part. :)

In particular, the part that was broken in my first draft was that I did not eagerly read errors from template args in types that
were NOT unique_ptr/vector/etc. -- the for loop at the end is actually very important for preserving the existing behavior, and
producing bindings that don't access private members etc.

The fact that there are now, in this version of the CL, no diffs, means this CL is OK!

PiperOrigin-RevId: 807747453
Change-Id: I06036081021c1e1c31e86b0c55c6059fcd5bd367