Check for fail-safe error types in template type parameters for vector/etc.
This is _needed_ for unique_ptr<Incomplete> outside of experimental, where, instead of being an IncompleteRecord, Incomplete becomes an Error. This causes compilation error for the bindings because we forget that it's incomplete.

But even aside from that, this fixes bad behavior with both types, where they generate Rust APIs like `unique_ptr<::forward_declare::Incomplete<...>>` for a type that doesn't have bindings, which is not correct even if it happens to compile.

---

It turns out there's no error checking for uniform repr types... oops! It can't go in the logical place, because after unknown commit, it now "depends" on which template we're implementing as to whether an error is acceptable. For example, if `std::allocator<int>` is an error type, we don't care. So, instead, I put it in the `type_arg` helper introduced in that CL.

More significantly, string_view is also such a type! It contains error types as the type parameters, but we don't care, because string_view is special-cased. So we need to add it as a special case here, similarly dodging the template parameter validation the same way vector and unique_ptr do.

I think this makes sense, anyway!

PiperOrigin-RevId: 807772718
Change-Id: I59f02d5108c14eeb9be58ffbb87186a064aefac9