Switch to the 2021 edition for the Cargo build.
I'd hoped to switch to the 2024 edition for the internal bazel build, but that turned out to be harder, because the toolchain is global, so you'd need to change every single rust_library and rust_binary individually to pass edition="2024".

Instead, probably the least effort to piggyback on the work others do.

---

As for why this says 2024: I think originally I had thought we'd already upgraded to 2024? IDK.

PiperOrigin-RevId: 811469433
Change-Id: I0e0e3ed515f9accd59a4a41ffde3545e935543d7