Generate C++ constructors for synthesized constructors of tuple structs.
Given a tuple struct containing zero or more public compatible fields,
we will generate a constructor that initializes all of them. If any of
those fields are:

  * Not public
  * Lack a C++ compatible type

We will omit the constructor and rely on the user providing a manual
`new`-like function to initialize. We will also omit the constructor
when the struct is marked `#[non_exhaustive]` because this forces the
synthesized constructor to be non-public.

PiperOrigin-RevId: 811867462
Change-Id: I1fce37808e5d9c378adb80ad756b56ad3c9fc9a0