Add a new Crubit flag, `non_unpin_ctor`
It is initially not used for anything, but just adding the flag to the set affects lots of golden tests.

This is the first half of a second attempt at https://github.com/google/crubit/commit/d46a71a56dd8929bf3057fc546c5328e75af0e36 after it was automatically rolled back for breaking a golden test.

PiperOrigin-RevId: 812826196
Change-Id: I498e5d50317501ff35bb2d37b80f1623213de639