Use a parsed format string instead of .replace() calls.
This matches the behavior of `format()` more closely, and lets us give errors if these aren't specified correctly. Also they're a little more efficient.

At, you know, the minor cost of an extra one or so hundred LOC.

The primary goal is to extract this to a library and use it to parameterize other strings, like the file path pretty-printing, in a followup CL.

---

Obvious question: why not reuse a library?

Actually, I was going to! But they seem to be unmaintained and have few users:

* https://crates.io/crates/rt-format - no updates in 3 years
* https://crates.io/crates/parse-format - code drop from 3 years ago
* https://crates.io/crates/runtime-format - no updates in 2 years

As far as I know we're generally discouraged from depending on unmaintained libraries. For example, https://opensource.google/documentation/reference/thirdparty/maintenance says:

> In order to maintain proper security posture in Google’s products, it is important that the third_party software at Google is actively maintained. [...]
>
> To minimize the risk that this poses, third_party packages which fail to respond to reports of security issues within 180 days will be locked down and considered deprecated. To prevent removal, owners and consumers of the package must either devise a plan to take on maintenance of the project and address the issues, or remove and delete the dependency.

This is a case where a dependency is arguably more technical debt than reinventing the wheel.

PiperOrigin-RevId: 813455953
Change-Id: If21a76e1255bb60e0a6b861ba84fa908ce727c5b