New query for local From impls by argument type.
Query provides a mapping from argument type (i.e. `MyCrubit` in `impl
From<MyCrubit> for i32`) to the local def id of the impl. This work can
be done once and shared between all ADT definitions for conversion
operator support, so I've added it as a new query. It is a relatively
narrow query. I think we could expand it to work for known traits of one
argument, but it wouldn't make sense to do this for queries of multiple
arguments. Without a concrete usecase, I've opted for the narrowest
functionality we need today.

PiperOrigin-RevId: 814804612
Change-Id: I00da85e406ede6c5182e36c6b15217492a1271ee