Make Crubit fail only during action execution when on an unsupported platform.
This is the most straightforward way I know of to make Crubit Just Work at analysis-time, and be ignored in an unused part of your dependency graph.

Alternatives I tried:

1. fail(): still executes at analysis time, doesn't allow deferring errors or anything.
2. return []: works, but leads to confusing errors of the form "unknown crate foo".
3. select() to avoid crubit deps in additional_rust_srcs_for_crubit_bindings: there does not seem to be a way to keep the select() and the toolchain definitions in sync, so I worry that this would inevitably either be too narrow (causing mysterious failures on supported platforms) or too broad (not resolving the bug).

This approach seems to me to be the one that's got the best error messages and is most "guaranteed correct" by virtue of actually reusing the toolchain selection logic.

The downside is a bit of extra complexity, and it takes much longer before a build error happens (it doesn't happen at analysis time).

This wasn't considered in the bug report (it was rejected option #4) because it was considered impossible. :D

PiperOrigin-RevId: 817716445
Change-Id: I9a20708fba4d36cc0b640a279267f740ea856910