Introduce function that summarizes the inference-relevant CFG portions and exports them in a protobuf.
To avoid code duplication, we refactor SummarizerAndCollector into a templated class that abstracts over the backend "collector" class. The new function then reuses the summarization functionality, but exports the summaries in a protobuf, rather than processing them direclty.

The existing functionality becomes a single standalone function which instantiates the template with the existing evidence collector class.

This patch includes definition of new protobuf messages to represent the summaries. A basic selection of tests is added to verify the protobuf messages are created as expected.

PiperOrigin-RevId: 817718410
Change-Id: I32b96843df6ab0d6118ace167cb310be985da968