Nullability inference: support overrides from previous inference iterations when processing summarized evidence.
In the phase2 of inference, SummaryEvidenceCollector can check for overrides of nullability annotations for the given symbol. The nullability of a param is thus adjusted during inference, and the new nullability is used for subsequent evidence collection steps.

The overlay only applies to "anonymous" declarations -- those that are not referenced by name and so not picked up in InferableSlots, nor modeled symbolically.

PiperOrigin-RevId: 817767923
Change-Id: I2ef9b536689ea5d31ca07977bf5bdb741bf9ba44