Test PR claiming we accept PRs
There is no proof like doing.

The problem with the other open PRs (oops!) is that they actually do touch lines affected by the copybara transform (that is the whole problem), so our copybara PR workflow won't work well on them. But this one should work! Probably!

Anyway, this might not work right away, but once it does work, the PR can be merged, and that's a win. :)