Unify bridge-names and existing-rust-type names.
I really regret not pushing for a generic "let `crate` stand for the current crate name" proposal! Then we would not need to parse as much of the structure. However, some parsing seems a bit inevitable.

With this change, both bridge types and types marked `CRUBIT_INTERNAL_RUST_TYPE` can be bridged/mapped to references, pointers, builtin types like `char`, and so on.

The use case for this is that in a followup CL I want to map proto enums to a crate-relative name using the existing rust type support, but it did not support crate-relative names, only bridging did. And I had to split it out to another change because, honestly, it was getting pretty intensive and really should be reviewed separately.

I've added some token unit tests, but the real test is in that it doesn't break our existing tests (and also, works in the followup CL.)

PiperOrigin-RevId: 818764980
Change-Id: I6d293c2d03d8c522d357fd3f570689c8c6910bcf