Support protobuf enums that are not nested inside of a message.
To make this work, I also generalized the path logic that's used by bridging to also work for ExistingRustType: if it begins with a `::`, it's treated as an absolute path, but otherwise, it's treated as being relative to the crate that defined the item.

This makes it so that we can special-case proto enums and treat them _as if_ they were generated with a `CRUBIT_INTERNAL_RUST_TYPE(ProtoEnum)` declaration.

(You might ask: why not use an enum item, instead of an ExistingRustType enum? The answer is: because then I can make them ABI-incompatible! Plus it seems more accurate, anyway.)

This doesn't work for enums nested inside a message because, well, given `message Foo_Bar {enum Bar_Baz {...}}`, we get a type like `Foo_Bar_Bar_Baz`, and there's no _easy_ way to get to `foo__bar::Bar_Baz` or similar. The enum is defined as the underscored name and aliased in the record, instead of the other way around.

PiperOrigin-RevId: 818767266
Change-Id: I13cbeb5b2cd511f85f67d481a24563ec245969db