Support protobuf enums inside of messages.
Still can't support top-level enums that have underscores in them. Don't do that.

This is about as far as I intend to take the bug, because top-level enums with underscores are really hard unless A) we traverse the whole header to look for the alias, B) convince protoc to emit an annotation.

So I guess I'm going to mark the bug as fixed, but people can reopen it if they really need this.

PiperOrigin-RevId: 818769375
Change-Id: I5dbb2fe906f760a7388783c3f07a9c29f5270845