Change C9Co to an RsTypeKind variant.
It's not a `UniformReprTemplateType` anymore, so this fixes that. I think it also doesn't generalize nicely into any of the existing variants. For example, it's not a bridge type because it's required to be at the top level, it can't be composed inside of something else.

In theory, it actually could be a bridge type, but in practice it cannot because we need to be able to write things like this:

```rust
trait CrubitAbi {
    const SIZE: usize;
}

fn foo<A: CrubitAbi>() {
    let _ = [0u8; A::SIZE];
}
```

PiperOrigin-RevId: 819847655
Change-Id: I799356501a49ea8e794cbdb7d64e33940629a44e