Change bridge.rs to be forward compatible for upcoming changes.
For example, the `TupleAbi` type is an unnecessary wrapper around a tuple, so I'm eliminating that. The more important change though is I'm setting up the codegen so I can transparently swap out the underlying behavior of `unstable_encode!` and `unstable_return!` to a more advanced approach where we allow `CrubitAbi`s to actually store context within them for the transformations to use, meaning codegen needs to be able to construct `CrubitAbi` values. This changes the codegen so that macros accept these values (and crubit generates their construction), but doesn't do anything with those values yet.

PiperOrigin-RevId: 823017533
Change-Id: I8bccbfef4103e664b1ffa4ae0919a7f8635162e7