Update logic to determine canonical name of a symbol.
Goal of this work is twofold:

1) Remove a usage of HIR and replace it with an rmeta-compatible mid
   usage.
2) Consolidate logic for determining canonical names

To that end this change attempts to maintain previous logic as much as
possible while changing the underlying implementation to work based off
of DefId's and ModChild (rather than HIR paths) and lays out that logic
in a sole source of truth `symbol_canonical_name`. There is more work we
could do, and we could pick better canonical names in some instances,
but both of those can build on top of this work once we're confident it
doesn't have unintended consequences.

PiperOrigin-RevId: 824723116
Change-Id: I16fed39f9bf671c7e51c37ad43711a844718f196