Fix bug around field naming bug.
With the recent change to `symbol_canonical_name`, we started
unkeywording more names than we did previously. This caused a collission
where our logic for unkeywording and for field/method name collision was
to suffix a `_`. We would unkeyword the method name to `operator_` but
also uncollied the field name to `operator_`. We now correctly handle
this case and uncollide the field name to `operator__`.

PiperOrigin-RevId: 825114473
Change-Id: Ib7531dd18f85d239aeac2878d3f9ee36f398055c