Do not collect attributed headers for targets that don't enable Crubit.
WTB: actual profiling tools for my starlark.

Since the number of targets that enable Crubit is very small, and will remain small for a long time, this was hypothesized to have an impact on build performance: instead of carrying around and communicating every single header from every single transitive dep (O(n)), we carry around only the ones that specifically enable Rust/Crubit support.

It does seem like a good idea on that grounds, and even if performance wins don't materialize here, it might be a part of the overall story later. That said, performance wins _don't_ materialize here, and the impact on bazel memory usage and CPU is indistinguishable from the null hypothesis.

Still, literally every CL in the chain to today, except for the immediately previous CL and this one, had a reason for existence other than the end goal of the bug, so there's only a little bit of wasted work. And it probably is marginally faster and better on memory, even if it's hard to measure.

Other effects: this does change the error messages: since we don't know what target a header is on, we can't say "enable Crubit on that target!", but instead say "enable Crubit on that header!". But the work necessary for this has all already been done: this is, as of this CL, _purely_ a change to the error comments, and not a particularly large one.

(The work necessary to unblock this was 1. to make it so that if a header is not known, it's treated the same as if it's known but Crubit was disabled on the target, and 2. to give error messages that mention the header file, in such an event.)

As of this CL, AFAIK, there is nothing left to do for the "shrink target_args" bug, and we can mark that as closed.

PiperOrigin-RevId: 825135672
Change-Id: I4cca69ac55410c28fdcc46bed5a08c283d62af48